+++ api/src/service/Client.h.ts
@@ -31,6 +31,30 @@ export interface BlockInfo {
   miner: string;
 }
 
+export interface PeerInfo {
+  id: number,
+  addr: string,
+  addrlocal: string,
+  services: string,
+  lastsend: number,
+  lastrecv: number,
+  bytessent: number,
+  bytesrecv: number,
+  conntime: number,
+  pingtime: number,
+  version: number,
+  subver: string,
+  handshakelocal: string,
+  handshake: string,
+  inbound: boolean,
+  startingheight: number,
+  banscore: number,
+  synced_headers: number,
+  synced_blocks: number,
+  inflight: Array<any>,
+  whitelisted: boolean
+}
+
 export interface BlockListItem extends BlockInfo {
   confirmations: number;
 }
@@ -109,6 +133,9 @@ export interface MultichainClient {
   // getinfo Returns general information about this node and blockchain
   // TODO add return types...although they seem rather flexible
   getInfo(): any;
+
+  // Return information about directly connected nodes (peers)
+  getPeerInfo(): Promise<PeerInfo[]>;
 
   ping(): any;